pyproject.toml
[project]
name = "kokomi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "jinja2>=3.1.6",
    "langchain>=1.2.15",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.12",
    "mcp>=1.27.0",
    "motor>=3.7.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "langchain-google-genai>=2.0.11",
    "httpx>=0.27.0",
    "uvicorn>=0.44.0",
    "langchain-community>=0.4.1",
]
